This update contains a couple of changes to the RTC code.
First, when a service is deleted from a node due to excessive downtime,
the node is no longer completely deleted from RTC - just the categories
based on that service. Bug 796.

Second, when SNMP is discovered on a device, the categories will refresh
to reflect any new rules the new SNMP data may have matched. Bug 683.

Created in this fix is a new method: rtcNodeIpRescan. Given a nodeid
and an IP address, this method will update the RTC information
for that key.

Needed: This rescan needs to happen when node data changes and when
assets are updated. Added an enhancement bug, 811.