Added a new util class to return a list of SnmpValues called getColumns. This allows one to walk a tree and return a list of values from that walk.
I then modified the SNMP monitor to accept the parameter "walk". Setting "walk"
to "true" will cause the SNMP monitor test to be applied to all of the OID values
returned from walking the base OID set in the monitor. If all pass, then the
monitor is UP, else it is DOWN.

This should be useful for walking things like a disk percent used tree, temperature
tables, etc.