Update the RTC categories whenever a node's surveillance category membership is changed via the web UI.  This way categories whose membership is wholly or partially defined by categoryName or catinc* rules will be updated in near real-time.  Fixes bug 2150.
Note that this changeset does *not* address the question of updating RTC categories when nodes' surveillance category memberships change as the result of a model import.  This case should be handled separately, probably by having the model importer tell Rtcd simply to reload everything, as that is potentially much cheaper than having Rtcd process a nodeCategoryMembershipUpdated event for every node in an import group.