More dependency injection of eventd -- who would have thought it!
There is now the concept of an EventProcessor (and an interface to go along with it)
and EventHandler now doesn't care about what actions it is taking on events, it just
iterates over the processor list, letting them each do what they need.  Here are the
processors, in order they are applied:
1) EventExpander: "expands" events based on eventconf.
2) JdbcEventWriter: persist the expanded event (if needed) to the database and sets the dbid on the event.
3) EventIpcBroadcastProcessor: calls EventIpcManager.broadcastNow to send the event to all interested listeners.
4) JdbcAlarmWriter: persist the expanded event (if needed) to the database as an alarm, either creating a new alarm or "reducing" into an existing alarm.

The EventProcessors are all thread-safe, and they are dependency-injected by Spring, so we can hopefully
eventually switch from the JDBC implementations to DAO-based implementations with our model objects and Hibernate.

Wheeeeee!