Approved by tarus/jeffg/david(?)@
a) Use a Resource instead of Files to reference the configuration information.
b) Support loading an event-file entry from a URL or a Spring psuedo-URL that
can be resolved by Spring's DefaultResourceLoader.  This includes "classpath:"
references (and there's a test for this, too). :-)
c) Convert it to be based on AbstractCastorConfigDao.
d) Change the classes that create DefaultEventConfDao to use the (now only)
empty constructor, property setting, and the afterPropertiesSet()
initialization method.  Do the same for bean definitions.  Also, pass
Resources, which makes some of the test stuff simpler than Files.