Modified two monitors:
SMTP: Changed the log messages to be more descriptive.

SNMP: Corrected the behavior of the "walk" option. It was supposed to require 
that all values of the walk passed the test, but due to a poorly placed if
statement only the last value had to pass.

Add a "matchall" option. By default this is true which means all values must
pass the test, just as before (well, now that the above bug was fixed). If
set to false, then only *one* value must pass the test. This makes it possible
to walk a table of, say, running processes and to insure that at least one
is running.