fix for bug #2634 (and a couple of related stuff I found)
This makes it clear just how much we need to redo the web UI.  85% of the code in
these JSPs are redundant accross them, and basically all of it is doing specialized
stuff it shouldn't.

I started down the line of refactoring the Filters to get rid of the duplicate code,
but oh man is that a maze.  Instead, I held my nose and committed this.  :)