This commit updates the SSH Monitor by renaming "protocolVersion" to just "version". We tend not to use camel caps in parameter names, and version is cleaner.
This commit updates the SSH Monitor by renaming "protocolVersion" to just "version".
We tend not to use camel caps in parameter names, and version is cleaner.

Added correct SSH and SSHv1 entries to capsd and poller configs.

TODO: Update the release notes to discuss new SSH behavior and mention that it is
easy to return to the old way.

Cleaned up my update to the SnmpMonitor by renaming "matchall" to "match-all".

Added an SNMP-based monitor to determine the status of a service in the Host Resources MIB
running services table.