BUG 2702
Define a 'Scope' for each PollEvent representing a Cause and to each PollableElement

Each pollEvent representing a cause has a 'scope' defined representing the scope of the outage.
This can be SERVICE, INTERFACE, or NODE and is determine from the UEI from the event.

Add Scope member to each element to represent the Scope an outage in this element represents

Inherit the cause of the parent when an element is created.
Make sure the null is handled properly when processingStatusChanges

Convert the PollerTest and PollablesTest to JUnit4 so its easier to debug in Eclipse

Revamp initialization from Database to properly determine causes for container object from members