Addresses enhancement bug 2864.
A new TrivialTimeMonitor for checking skew of a remote clock using the
trivial UNIX time service.  This is a rusty old shotgun compared to the
sniper rifle that will be the finished NtpMonitor, but still useful.
Configuration examples are in etc/examples only since this is targeted
to a stable point release.