remote poller startup behind restrictive firewall
As a side-effect of the way the remote monitor jar is generated, it
attempts to look up XML schema definitions online, which does not
work if the host is behind a firewall or in any other kind of
restrictive environment.

This patch changes the remote poller to use One-Jar
(http://one-jar.sourceforge.net/) to create a remote-poller jar which
contains all of the component jars as individual pieces, which fixes
the problem of various resources getting overwritten in the old
unpacked assembly style jar.

Fixes bugzilla #2827 and Jira PJSM-3, PJSM-8, and PJSM-9.