HTTP unit test framework
This is the basic structure for an HTTP unit test framework.

To use it, create a spring-based JUnit test with the
JUnitHttpServerExecutionListener added as an execution listener.

Then, annotate your class and/or methods with the @JUnitHttpServer
annotation.

By default, it will serve up any files in target/test-classes, on
port 9162.  For more details, see the
JUnitHttpServerExecutionListener class javadoc.