bug #3520 - create datalink with proper status
Originally, the link adapter would only create datalinks with
a status of "G" regardless of whether the endpoints were up
at the time of creation.  This could cause a condition where
the state of the link would never change until a service were
to go down and then come back up.