Enable Spring Security on the Remoting Webapp
This patch enables Spring Security on the remoting webapp, as well
as a whole tonne (British!) of other refactoring to make it possible
to do so without pulling in tons of code needlessly.

* move some event utility code into opennms-model
* create features/springframework-security to hold the "common"
  OpenNMS spring security code (moved from opennms-webapp)
* create opennms-web-api to hold web API code that is shared
  between the opennms webapp and remoting webapp
* create tests/mock-elements to hold a number of the mock classes
  used by unit tests to avoid pulling opennms-services into some
  of the lower-level stuff

There's more to do with tests/mock-elements, but this is enough to
make things work.

I've added a role ("remoting") to magic-users.properties which can
be used for connecting the remote poller to the HTTP poller backend.
It will also accept connections as administrator users.