CBNL CORE: Fix send-event.pl script to encode time in DateFormat.LONG style
The core org.opennms.netmgt.xml.event.Event object that maps events
stores time as a Java DateFormat.LONG string. If the string placed
here doesn't meet the format parameters this can cause exceptions to
be thrown later when the string is parsed into a Date object
(hopefully using the EventConstants.parseToDate method).

While this fixes this problem it would probably be better if the event
threw an exception when the bad date format was set rather than
waiting until someone tries to process the information.