Initial work to fix Bug 3554
First, ThresholdingSet.applyThresholds code was updated to skip
processing if there are no valid attributes, because it not make
sense.

After a minor refactoring to ThresholdingVisitorTest, two junit tests
were added about Bug 3554 with the config files attached to that bug.
One with a mock FilterDao and other with a JdbcFilterDao. Both tests
works, and the second does not appear to take to much time.

Still need more research for Bug 3554.