Enable OProfile profiling support
This patch to the opennms start up script allow a command line flag to
be passed that will load the OProfile JVMTI module. This module will
allow OProfile to show more detail as to where hot-spots are than just
the big chunk of anonymous memory allocated for Java's JIT code.