Fixed bug #3620
An operator is aware if an interface is snmppolled or not.
The changes are the following:
a) modified snmpinterface table to support snmppoll (P means Polled, N means
Not Polled) and snmplaspsnmppoll (a Date) fields.
b) Enhanced OnmsSnmpInterface in the Opennms Object Model to support these new properties
c) refactored the "SnmpPoller" service to mantain the snmppoll and lastsnmppoll fields
d) Changed the gui to have the snmp poll status date and package name showed
   either in element/snmpinterface.jsp either in element/interface.jsp pages
e) refactored the configuration file according with the not deprecated configuration
   properties