Fix for Bug 3427 (1.6 branch only)
All ThresholdEvaluator classes were updated to format values with
java.text.DecimalFormat, based on the system property:

org.opennms.threshd.value.decimalformat.

The default value for this property is '###.##', which means that the
threshold value will show only 2 decimals. If you want to override
this, add the property to the opennms.properties file.