Fixed snmp-graph.properties to be 100% compatible with RRDtool
The rrd_graph documentation said:

"A colon ':' in a legend argument will mark the end of the legend. To
enter a ':' as part of a legend, the colon must be escaped with a back-
slash '\:'.  Beware that many environments process backslashes them-
selves, so it may be necessary to write two backslashes in order to one
being passed onto rrd_graph."

I used a simple perl script to automate the process:

while (<>) {
    if (m/\w+:.*"[^":\\]*:/) {
        s/"([^":\\]*):/"$1\\\\:/;
    }
    print;
}