Refactor the location manager.
This refactors the location manager so that it contains
all of the behavior specific to a given implementation of
a maps API.  This includes changing the DefaultLocationManager
to the GoogleMapsLocationManager, and moving a bunch of
Google Maps API calls from the Application class into the
GoogleMapsLocationManager.

It also optimizes communication between the backend and
frontend to push locations as an atomic group, and changes
the UI refresh to not occur until everything has been
updated.