refactored packaging for the remote poller
This patch makes a new separate project to house the remote
poller "standalone" distribution, and changes the base assembly
to include it properly (in bin and contrib).

It also changes the debian and RPM packages to pack them up
alongside init scripts to let them start up automatically on
boot.