Last Commit for  Bug 3751: Refactor SVG Context Menu
Created a context menu for the maps....when you double click on a map you got a context menu that enables
the ability to navigate throw the maps hierarchy.
The menu' shown the list of parents map + open option.
Clicking on the parent will open the selected parent map. Clicking on open will open the double clicked map.