Fixing Bug #2910
As the suggested formula for the threshold memory works only on Linux
systems, two threshold packages were created: one for linux, and one for
non-linux systems.

The package for linux has a threshold group with the formula suggested.

The package for non-linux has a threshold group with the original
formula. This formula has been removed from netsnmp group.

The result will be that the suggested formula will be applied only on
Linux systems.