Refactoring datacollection-config.xsd
The content of datacollection-config.xml will be exactly the same as
other versions of OpenNMS (to avoid problems with upgrades, and with all
users which want to use the old style).

But, now you can use the new attribute "include-collection" to make
references to resources from external files.

The parsing of the splitted data is not available. It will be soon!.