Updated DataCollectionConfigParser for resourceType handling.
All resource types defined on datacollection-group's files, should be
added to each snmp-collection object, to ensure that the
DataCollectionConfig implementation has all resource types (used or
not).

DefaultDataCollectionConfigDaoTest has been refactored, to validate that
the old datacollection-config.xml and the new one, produce the same
content from the DataCollectionConfig interface point of view.