Fix for Bug 4117
The problem with IfLabel and Vault, is that Vault.setDataSource is
called only on ServletInitializer, which always will run after Pollerd
is up and running.

For that reason, vault requires obtain a DataSource by itself using,
for example, DataSourceFactory.

A calling to DataSourceFactory.init is not required, because that is the
purpose of DataSourceFactoryBean, which is initialized by Spring on
applicationContext-dao.xml, and that is done 'before' Pollerd is up and
running.

So now, when Vault obtains the DataSource using the Factory, it won't be
Null anymore.

I tested it and it works. I can see the events and alarms related to
latency thresholding processing.