Major rework of the build system scripts.
This replaces "build.sh" with 2 scripts, compile.pl
and assemble.pl.  It is now possible to re-assemble the
tarball that used to be generated with "assembly:attached"
or "assembly:directory-inline" without having to re-build
other parts of the OpenNMS source tree.

For info on how to use the new commands, see:

  http://www.opennms.org/wiki/Building_OpenNMS#Building

Also, check the output of "./compile.pl --help" or
"./assemble.pl --help" for more details.