NMS-4307 - open support tickets from the UI
The help/ path in the web UI has been refactored to "support".  The
support index has been changed so that it includes links to a web UI
for creating system reports to aid in debugging, as well as a form
for opening a support ticket directly in the OpenNMS.com ticketing
system.