Making RrdDao happy with RRDtool 1.4.x.
When you execute a 'rrdtool graph' command in order to retrieve data
using PRINT statements (like JniRrdStrategy), if the DS does not
contains any data for the required time range, it will return the word
'nan' for RRDtool versions 1.2.x or 1.3.x. For 1.4.x it will return the
word '-nan'.

This FIX updates the code to handle both sentences ('nan' and '-nan').

Without this fix, Statsd and RrdGraphController won't work with RRDtool
1.4.x.