Fix for Bug NMS-3905 ~ Reload threshd-configuration.xml
[Merged from 1.10 to 1.8. Includes the changes on opennms.properties]

Removed the global property org.opennms.thresholds.filtersReloadEnabled
and the code related to execute rebuildPackageIpListMap from the
ThresholdingSet because that should never happen. The correct way is to
call the reload method from the ThreshdConfigFactory (which will call
rebuildPackageIpListMap), when the event named reloadDaemonConfig is
processed for threshd-configuration.xml. This doesn’t have any problem
with NMS-3599.

Refactored the initialization of the ThresholdingVisitor to only return
null when the threhsolding is disabled. If it is enabled but there is no
thresholds defined for a specific visitor, the new method called
hasThresholds will return false. This method is used before call the
visitor from CollectableService.

Added a new JUnit test to validate the reload of the file
theshd-configuration.xml

Added a button on the WebUI to reload theshd-configuration.xml