Add validations over the services count for HostResourceSwRunMonitor
A JUnit test has been added to prove that all combinations work as
expected.

If "match-all" is true, there can be an optional "min-services" and
"max-services" parameters that can define a range. The service is up if:
a) services_count >= min-services and services_count <= max-services
b) either one is not defined, then only one has to pass.
c) neither are defined, the monitor acts just like it used to - checking
all instances to see if they are all running.
It is assumed that all services would have to pass the minimum run state
test, no matter what the count.

I also added some additional features already implemented 1.10 for this
particular monitor to enhance their usability.