SPC-25 - fix alarm/hibernate implementation
Re-did the changes related to SPC-25, including changing from
an EnumType.ORDINAL to a UserType, since enums are 0-indexed
in Hibernate, and we want to start at 1 (to be more specific,
we want to use whatever getId() returns, rather than just the
index in values()).

Also, a few other related cleanups.