Always storing all records in the database no matter if status=0.
"The timesync status 0 indicates that device is not locked to the master
clock. What it means is that the relative clock offset between the local
device and remote device (grandmaster) could be larger than the
constraint value, ex. 10 microsecond. It doesn't necessarily conclude
the data is bad. If the delay is in the millisecond range which will be
most likely the case in the real world mobile backhaul network
environment, the microsecond (ex. <50 microsecond) clock offset is ok
for millisecond delay measurement."