Add a new generic "Criteria" query object
This adds a new "Criteria" object and builder which encapsulates
operations to be done when querying data from the database.  It
lets us have a more easily-manipulatable and reusable builder or
object for making queries to the database, and keeps us from
having to do funky state-manipulation in the old OnmsCriteria
APIs.