Changed ConnectionFactory classes to reuse localhost ports with the SO_REUSEADDR option set to avoid overloading the system with lots of sockets waiting in the TIME_WAIT state after the connection closes. This should fix bug NMS-5469 which was introduced after my last round of ConnectionFactory changes.
Changed the check for missing banners so that both sides of the
connection must be idle before we consider the banner missing. This
should avoid the problem in bug NMS-5311 which only occurs under heavy
load.

Increased the number of iterations inside
AsyncDetectorFileDescriptorLeakTest from 10,000 to 30,000 since it was
taking more than 10,000 to see some high-load problems on my test
machine.