Integrate OSGi servlets into the webapp's
This commit merges the features/osgi-servlet-fix branch, which adds a servlet
filter that sits in front of the normal servlet dispatcher and checks if a
URL is mapped to a servlet inside the OSGi bundle context (through Felix's
servlet bridge & proxy).  If it is, it will call into that servlet, instead
of letting our existing webapp servlet dispatcher resolve it.

This means that any servlets registered in OSGi will be accessible directly
through their mappings without needing a special /opennms/osgi path or
anything like that.  For example the topology servlet can now be reached
directly at:

	http://localhost:8980/opennms/topology

These changes include a slightly modified version of the felix bridge and
base code to facilitate using it as a servlet Filter, rather than an
HttpServlet.