Removing mib2events dependency and minor fixes.
Because mib2events is using lots of unnecesary dependencies, I ended up
by copying the important code from mib2events to MibbleMibParser to make
it work, and now everything is working better in Karaf.

I've added a handler to remove MIBs from the pending directory and I've
fixed some minor issues when detecting if the files already exist or not
when uploading a new file.