Be able to recognize -inf/inf on RRD's XMLs and updating rollback procedure.
Instead of doing a full rollback when a single interface was not able to
be merged, the code has been modified to avoid a full rollback. Instead,
if an interface was not able to be merged, only that interface will be
rolledback. That way, we can continue with the upgrade process.