Place the upgrade status file on $OPENNMS_HOME/etc
Originally, the tool was creating a hidden file on $OPENNMS_HOME/bin. In
order to be able to avoid problems during upgrades (especially because
backing up the etc directory is part of the usual procedure), the file
has been moved into etc directory and it is not hidden anymore.

The filename is: opennms-upgrade-status.properties

Also, the installer has been updated in order to be able to pass a flag
to skip the execution of the upgrade tools. This can be useful in some
occasions, specially if the installer is embedded into a big custom
installer script (like the Juniper Integration):

 -S,--skip-upgrade-tools        Skip the execution of the upgrade tools
                                (post-processing tasks)