Feature: Upgrade Tools API
This new API provides the ability to perform post-processing tasks after
upgrading OpenNMS, that can be execute with OpenNMS up and running or
with OpenNMS stopped.

By default the Installer class has been modified in order to execute the
Upgrade Tools after its completion, unless the user specify a flag (-S,
--skip-upgrade-tools) from the command line to skip their execution.

There are currently 2 implementations enabled:

- Remove non-ip-snmp-primary and non-ip-interfaces from requisitions:
  NMS-5630, NMS-5571 (Offline)
- Merge SNMP Interface directories (Online Version): NMS-6056 (It works
  with RRDtool and JRobin, and with or without storeByGroup enabled)

There is one implementation that it is currently disabled until the
required code changes are added:

- Fix the JRB names for the new JMX Collector: NMS-1539, NMS-3485,
  NMS-4592, NMS-4612, NMS-5247, NMS-5279, NMS-5824

The API basically requires to implement an interface called OnmsUpgrade.
The Upgrade Tools are going to find all the implementations of that
interface inside the package namespace "org.opennms" , that doesn't have
an @Ignore annotation. Then, all the implmentations are going to be
ordered, and they are going to be executed according with the current
OpenNMS running status, after checking that they haven't been executed
before.

The execution status is going to be stored on
$OPENNMS_HOME/etc/opennms-upgrade-status.properties.

The implementation also includes a new package called
"opennms-rrd-model" that contains all necesary classes to parse the Dump
of a JRobin or RRDtool file, and the output of an RRDtool Xport command.
Besides that, there are tools to convert from JRobin to RRDtool, to
split one multi-ds file into several single-ds files, and to merge a
list of single-ds files into one multi-ds file. The object
representation is handled thorugh JAXB. In the future, this tools can be
used to create ReST entries for raw or processed data from JRobin or
RRDtool files.