Refactoring the upgrade tools to avoid RrdUtils.
Now, AbstractOnmsUpgrade use similar algorithm like JRobinDirectoryUtil
to determinate which RRD strategy is enabled and the configured RRD
extension without depending and initializing RrdUtils.

The benefit for this is to avoid the ClassNotFound (related with
JniRrdStrategy) thrown when running the install script with RRDtool
enabled, because RrdUtils uses Spring to initialize the RRD Strategy and
that conflicts with how the class loader has been initialized on the
Installer.

Also, the dependency to core/upgrade has been removed from
opennms-install because the upgrade tools are now being instantiated
like this:

System.setProperty("opennms.manager.class", "org.opennms.upgrade.support.Upgrade");
Bootstrap.main(new String[] {});