Forcing some CSS styles on the <pre> tag used to show errors.
The MIB Compiler builds the errors when they exist with a specific
format. For this reason, the MibCompilerPanel uses a <pre> tag to
display the content. The problem is that when the content is too big,
it is going to be truncated, so the idea of the styles is to wrap the
content accordingly (compatible with all kind of browsers).