Cleanup vaadin code for managing events and datacollection (VI)
1. New Validations:

a) If a resourceType is going to be renamed or deleted and it is being
referenced by a group, the operation will be aborted.

b) If a group is going to be renamed or deleted and it is being
referenced by a systemDef, the operation will be aborted.

c) If a systemDef is going to be renamed or deleted and it is being
referenced by a SNMP collection, the operation will be aborted.

d) If a SNMP collection is going to be renamed or deleted, a warning
message is going to be shown to alert the user that Collectd's
configuration might require an update.

2. Fixing initialization of IncludeCollectionWindow.