Fix for Issue NMS-6386
The ILR reports wrong data when there are several packages with
different collection rates on collectd-configuration.xml for the same
service.

Now, all the operations are performed in CollectableService and
CollectionSpecification. There is no need to add additional entries on
the collectors implementations.

The CollectionErrors or CollectionUnsuccessful are now properly recorded
as "errors" in terms of the ILR report (which was only working for SNMP
before this commit).