Fix for Bug NMS-6484
The Valere devices with broken SNMP agents are hanging Provisiond.

A defensive code has been added to abort the walker execution if a valid
PDU (i.e. no errors) with 0 varbinds is returned by the agent, which is
an indication of a broken agent.

The logs are going to show a warning with the following content:

Aborting node scan : Agent failed while scanning the system table: Error
retrieving systemGroup for /X.X.X.X: A PDU with no errors and 0
varbinds was returned from the agent at /X.X.X.X. This seems to be
related with a broken SNMP agent.