Fix for Bug NMS-6274
Thresholding ignores persistselectorstrategy resulting in threshold
events on items that should be ignored.

Threshold processing is not going to be performed if the output of
CollectionResource.shouldPersist(ServiceParameters) is false.

That way, the same rules defined to decide if a CollectionResource must
be persisted or not, are going to be used to decide if we should apply
or not threshold processing on the resource.