Fix for Bug NMS-6545
ResourceDao.findTopLevelResources() is returning duplicated data because
of storeByForeignSource.

Tested on the following scenarios:

1) All nodes are provisioned and storeByForeignSource is disabled (i.e.
all the RRD directories are like this: rrd/snmp/{nodeId}.
2) All nodes are provisioned and storeByForeignSource is enabled (i.e.
all the RRD directories are like this: rrd/snmp/fs/{fs}/{fid}
3) Hybrid environment, there are requisitioned nodes and discovered
nodes on the same installation with storeByForeignSource enabled. In
this case, the discovered nodes will have rrd/snmp/{nodeId} for the
RRDs, while the provisioned nodes will have rrd/snmp/fs

IMPORTANT: If storeByForeignSource is enabled and the RRDs are stored
under a directory like rrd/snmp/{nodeId}, those RRDs are going to be
ignored (because they must be stored under rrd/snmp/fs/{fs}/{fid})