PJSM-212 Fix embedded XML in poller config (PSM)
When the Poller configuration was converted to JAXB, the config object
now contains the actual objects embedded inside parameters, instead of
an XML string that had to be hand-parsed.  However, the code inside
PollableServiceConfig was calling .toString() on the embedded config.
So, for example, an embedded page sequence would actually be returned
to the PSM as the string representation "PageSequence [Page=foo]" and
it would fail to monitor.  This likely affected other embedded XML like
the MTM as well.

Squashed commit of the following:

commit b7e90c1ee2bd9c786eff45d8d0694336d80d0458
Author: Benjamin Reed <ranger@opennms.org>
Date:   Mon Jul 14 20:35:20 2014 -0400

    better mocking; move to a separate test

commit ca1679f59b5af63fa8d39d877ddb4e599a483542
Author: Benjamin Reed <ranger@opennms.org>
Date:   Mon Jul 14 16:52:30 2014 -0400

    get rid of extraneous field in the test

commit 6da4ce1fe063729ed836268cd057e4ca750f376a
Author: Benjamin Reed <ranger@opennms.org>
Date:   Mon Jul 14 16:34:35 2014 -0400

    PJSM-212 fix handling of objects passed as service parameters

commit 8f9ba982b0874cf712ab0bc0bd16565165eda401
Author: Benjamin Reed <ranger@opennms.org>
Date:   Mon Jul 14 16:15:15 2014 -0400

    remove old useMoxy code, we use it by default now