Fixes related to unexpected self-reverting behavior when users modify the asset records and node category memberships of requisitioned nodes.
For the "Edit Category Memberships" flow for a single node, added a
warning blurb plus a checkbox that the user must tick before being allowed
to submit the form.

For bulk edits of category memberships from the top-level Admin menu, put
requisitioned nodes in separate option-groups inside the select boxes,
initially in a disabled state, and added a checkbox which when ticked will
enable those elements.

For single-node asset editing flow, which is GWT and harder to modify, did
a good-enough speed bump in the form of a Javascript confirm dialog which
is shown only when the node is requisitioned AND the user is not in the
read-only role.

Squashed commit of the following:

commit 9e518d735825805cebbf8e68161f8aff3b0c7153
Author: Jeff Gehlbach <jeffg@opennms.org>
Date:   Mon Jul 14 10:18:56 2014 -0400

    Do not show asset-editing speed bump for users in the read-only role.

commit d0ba517f75a418db435ff0c258ec9d240ad5eae1
Author: Jeff Gehlbach <jeffg@opennms.org>
Date:   Fri Jul 11 10:56:53 2014 -0400

    Speed bump for asset changes on requisitioned nodes

commit 87d785a6e4535df0e39e0fa13e7f955c8633f7b3
Merge: 339842c 8c0dc3a
Author: Jeff Gehlbach <jeffg@opennms.org>
Date:   Fri Jul 11 09:59:15 2014 -0400

    Merge branch '1.12' of github.com:OpenNMS/opennms into features/category-asset-requisition-unfubar

commit 339842ce45f379eca82736176b705f91fd2831b9
Author: Jeff Gehlbach <jeffg@opennms.org>
Date:   Wed Jul 2 16:41:22 2014 -0400

    Speed bump for mass category changes on requisitioned nodes

commit aea95082e48ef00ad3164129301bbe7ea78677e0
Merge: 2e55e06 ae606f9
Author: Jeff Gehlbach <jeffg@opennms.org>
Date:   Wed Jul 2 11:08:14 2014 -0400

    Merge branch '1.12' of github.com:OpenNMS/opennms into features/category-asset-requisition-unfubar

commit 2e55e067e5f77a9d583b098983725f9716243b40
Author: Jeff Gehlbach <jeffg@opennms.org>
Date:   Thu Jun 26 17:33:14 2014 -0400

    Speed bump for category editing, road closed for node label changes on requisitioned nodes