Fix for Bug NMS-5219
Threshold expressions for node-level metrics doesn't work on non SNMP
Collectors.

The code for the XmlCollector, JdbcCollector and WmiCollector has been
updated in order to ensure that only one node-level resource is going to
be created per collectionSet.

Note: the HttpCollector must be updated as well, but considering the
complexity of the current implementation, I've decided to leave it
untouched. Avoid creating expression based thresholds for this
collector.