Fixing the SQL Schema and the Hibernate relationship between OnmsNode and OnmsHwEntity
Now, there is nothing inside OnmsNode (or the node table) related with
the entity. It is required to use HwEntityDao to retrieve the root
entity of a given node by its ID.

The ReST API for hardware entities has been fixed to read and write root
entities for a given node.

NOTE: In order to replace or update the hardware inventory of a given
OnmsNode, the code will remove from the database the current content of
the hardware inventory if exist, and then it will be replaced with the
provided data.