Improving the hardware adapter to reduce overhead.
The adapter is going to initialize the list of available vendor
attributes at the beginning and when the configuration is changed.

The SNMP Plugin analyzes the HwExtensions to obtain the list of OIDs for
the node it is going to process.

Pending: implement the Package on the configuration to filter nodes.