Using Orika for merging change after re-synchronizing an existing node.
The old code was deleting the existing hardware information, and adding
a new one; but that genering some Hibernate/PostgreSQL exceptions for
the defined constraints.

For this reason, I decided to merge the collected entities with the
existing entities in case they exist on the database. Merging beans is a
very complex operation, and that is why I decided to use Orika for this
purpose.