Applying some fixes to the OnmsHwEntity and related classes.
After parsing the XML representation of the hierarchical structure, it
is important to fix the internal relationships in order to be able to
properly store the structure on the database.

The JUnit test for saving hardware inventory has been updated in order
to validate that the XML representation can be stored successfully on
the database.