Adding required insfrastructure to select requisition strategy.
A factory class has been created in order to dynamically select the
strategy used to manage requisition based on the value of a system
property called org.opennms.provisiond.repositoryImplementation,
declared on opennms.properties. The default value is file (the currently
used strategy).

Available strategies:
- fastQueueing
- queueing
- fastCaching
- caching
- fastFused
- fused
- fastFile
- file (default)

The default qualifier (i.e. "deployed", "pending") for the
ForeignSourceRepository interface are going to use the factory
to obtain the configured strategy.