Re-implementing DirectoryWatcher using Java7's WatchService.
Because an infinite loop waiting for file changes is required, the
implementaion is going to create and maintain an internal thread for
handling the changes on the requisitions.

Tested on a live OpenNMS with more than 500 requisitions and more than
10000 nodes in total. The response time of the Requisitions ReST API is
almost instantaneously. Of course a cache of the requisitions is kept on
memory in order to accelerate the response time.

Pending work: fix the logging for DirectoryWatcher, because the logs are
appearing on eventd.log.