Adding defensive code to avoid NPE
I'm not 100% sure why the NPE happens, I guess it is because of a race
condition due to the fact that each DirectoryWatcher instance is going
to create and maintain a separate thread for looking at directory
changes.

So, I've added a NPE check on the FileReloadCallback implementations
defined on FasterFilesystemForeignSourceRepository.