Fix for NMS-7230 Part II
Refactoring WallboardProvider to avoid keeping a reference of a
BeanItemContainer, to aliviate the memory problems.

Doing a more deep analysis of the heap dumps, the leak suspects were
pointing to the WallboardProvider. This class is responsible for
providing the wallboard configuration, so the new implementation is
focusing on that, and now it creates a new BeanItemContainer class
each time the client side make a request for it.