Fix for enhancement NMS-7438
Apply the auto-resize feature for the timeline charts.

The logic for resize resides on js/timeline-resize.js. It is going to
use 50% of the panel size for now due to the restrictions I've mentioned
on Jira.

In case a change is required, only that file must be modified instead of
3 separate JSPs.