Improving and verifying the merge function on opennms-rrd-model.
Now, it is possible to merge multiple single-metric RRDs to a single
multi-mertic RRD; which provides the ability to convert from
storeByGroup=false to storeByGroup=true and any combination of
JRobin/RRDtool.

Also, the split version should work on the opposite way, in case someone
wants to migrete from storeByGroup=true to storeByGroup=false.

The tests on the XML Collector for RRDtool has been updated to be sure
they work. Also the opennms-rrd-converter has been updated to use the
opennms-rrd-model. This will improve the performance of this tool
drastically.