Creating AbstractSpringJerseyRestJsonTestCase.
This class extends AbstractSpringJerseyRestTestCase to force use JSON
instead of XML to test the ReST API.

The idea behind this is to make sure that the ReST API supports JSON.

I started with RequisitionRestServiceJsonIT. It does the same thing as
RequisitionRestServiceIT but with JSON. Instead of doing a simple string
compare like the XML version (which I think should be refactored to
parse the XML), it uses org.json.JSONObject to parse the output and
build the requests, to ensure that the format is consistent.

Also, some fixes has been added RequisitionRestServiceIT to properly
validate the XML output.