Simplifing the WatcherService thread.
It is required only to track modifications and deletions. The action
should be just remove the entry from the cache. On next attempt, it will
be loaded.

Unfortunately the same use case that has been failing it is still
failing:

1) Add a requisition with one node
2) Synchronize the requisition
3) Add a second node
4) Synchronize the requisition

At this point if you follow the redirects, you'll receive the correct
content of the requisition, but when Provisiond starts the import, the
requisition is overriden with the old content, and the new node is not
added to the database.