HZN-344: First version of metric-converter running
Reworked the metric crawling including a bunch of bug-fixes for path
handling and configuration detection.

It now also supports converting installations where storeByGroup is off
and using the group from .meta files

Insertion of string properties was added.

TTL handling was fixed, so the remaining TTL is relative to the current
time.