Fix for enhancement HZN-563 - Part I
Initial implementation of the SyslogFilter.
The core fields from OnmsNode (nodeLabel, foreignSource, foreignId and
sysObjectId) will be available in the NorthboundAlarm. Because of this,
there is no need to depend on NodeDao when extending AbstractNorthbounder.
The NorthboundAlarm class was modified to provide a list of Event Parm
and a string map with all the event parameters.
A simple test has been added for SyslogFilter.

Pending:

Create JUnit test to verify the proper behavior of the Syslog
northbounder with filters.
Implement the auto-reload of SyslogNorthbounderConfigDao when the
configuration file is changed.
I suggest to properly implement reloadDaemonConfig for the northbounders
through Alarmd, and implement a method called reload to the Northbounder
interface.