Fix for Enhancement HZN-564
Create an alarm northbound implementation for SNMP Traps.

All the basic functionality has been tested through JUnit tests.

Pending:
Create JUnit test to verify the registration/unregistration of the
northbounders to be sure they work.

Possible Enhancement:
Check if the content of a trap sink was changed, and only re-register
it, if the content was changed.