HZN-584: Initialize WS-Man documentation from draft
Build structure for performance data collection. The structure follows the same pattern as in Service Assurance. The
content is migrated from the Google Docs draft file to Asciidoc format. The description for the WS-Man detector is
separated in a dedicated file. There is a good chance this section will move to Provisioning detectors.