NMS-8113: Whats new section in release notes
Introduced Whats New section for major releases and have changelogs for minor releases. Removed requirements and known
issues for the reason they are outdated and not maintained. Requirements should be covered in the install guide of
OpenNMS. Updated copyright dates to 2016.

Fixed wrong syntax for anchors which can be used as internal reference.