Upgrading AngularJS from 1.12.28 to 1.4.9
Also Angular UI Bootstrap has been upgraded. The API for this library
was changed but it works better than the old one. The hack for the
typeahead is not needed anymore. Now it works as expected when you want
to show all the options as soon as the input field is on focus.

All the karma tests are passing. All the features have been tested on a
running environment.