Major refactor of the AngularJS Service for Requisitions.
Two ReST end-points have been implemented:

- /rest/requisitions/deployed/stats
- /rest/requisitions/deployed/stats/{foreignSource}

The above end-points provide the requisitions that have been updated
with the following information:

- Name of the requisition.
- Last deployed time.
- List of deployed foreignIDs on the database.

The UI provides a refresh button to retrieve the deployed statistics and
update the current data model on demand, without the need to reload all
the requisitions.

The load process is a lot more efficient now, because the deployed
requisitions are not required anymore.

A few visual enhancements has been added to let the user know when the
requisitions have changes that require attention.