BSM-148: Add information panel to Topology UI
This provides the infrastructure and APIs for selection responsive info
panels in the topology UI. Registered panel items can provied UI
components in relation to the selected vertex or edge.

A detailed panel for all BSM components is provided showing the element
details and the status including an inside view of the status
calculations.

In addition, a generic panel for technical details is provieded for all
selections, regardless of the specific type.