Manually revert of Pull Request 671 (NMS-8253) based on current state.
I found that the solution is not going to work as expected due to how
the Path Outage feature is implemented, and how the ServiceMonitor
implementations are handled by Pollerd.

The current code will be like the original one without the need of
having IcmpPlugin. It contains a JUnit test to verify the behavior.