BSM-137: Add indicator if in simulation mode
This required to refactor the current Info Panel API.
It is now possible to provide content based on the GraphContainer's state and not only based on the selection.

This enables Topology Providers to provide static and dynamic content.

In this scenario a 'WARNING Simulation Mode' message is shown in the Info Panel if in simulation mode.