NMS-8043: Revert deletion of ExecCommandController
When linkd was removed from OpenNMS the ExecCommandController and ExexCommandAjaxController were also removed (see https://github.com/OpenNMS/opennms/commit/6191861271648cb1424d47fe50a69d1b310cc60d).
The ExecCommandAjaxController was only used by linkd code.
However the ExecCommandController also handled the ping and traceroute operations from the TopologyUI.
Removing this class results in a 404 Page Not Found error when invoking the ping and traceroute operations (from the Topology UI).

In order to still be able to use the existing functionality the ExecCommandController deletion is reverted.
However in general it is strongly adived not to use this functionality in the future.
See http://issues.opennms.org/browse/NMS-8245 for why and more details.