Fix for Enhancement NMS-8484 - Part V
Adding support for Hierarchical Cisco QoS Class Map Statistics, by
implementing CiscoQoSPropertyExtender.

This requires to define the data collection like shown in
datacollection-config-cisco-qos.xml; the configuration used
for the JUnit test.

All the aliases and resource types for the required fields can be
changed through property parameters, in case a user is already
collecting the data.