NMS-8704: Add an IBM Tivoli Event Integration Facility adapter to OpenNMS. The EIF adapter accepts events sent using the EIF protocol from IBM Tivoli systems including ITM and Netcool/OMNIbus, and parses them into OpenNMS events.
NMS-8704: Project layout and initial routing code for the EIF adapter.

NMS-8704: Add EIF parser and accompanying unit tests.
Rewire OSGi blueprint to use Event.Log instead of Event, so we can return multiple events.

NMS-8704: Use a custom ChannelHandler instead of a Processor so we can handle multiple clients in parallel.

NMS-8704: Add a Logger for logging messages.
EIF event data starts 37 chars after the <START>> tag. Use that instead of a smaller offset and a regex to find the event data.
We still use a regex to identify the EIF event class at the beginning of the data, but it's much less delicate now.
Add a NodeDao bean so we can look up node data.
Add an enum to translate EIF severities to OpenNMS severities.
Use the slots 'fqhostname', 'hostname', and 'origin', in that order, to attempt to look up the event's node. If all else fails, pass nodeId0 and log a warning.
Use a more standard UEI, and update the unit tests accordingly.

NMS-8704: Wire the NodeDao into the OSGi context and message decoder, and pass it to the parser.

NMS-8704: Refactor EifParser and tests.

Refactor EifParser:
 - Remove unneeded regex for event body extraction
 - Terminate the EIF event with the eventEnd index instead of (eventEnd - eventStart).
 - Simplify the node lookup code by first trying all 3 strategies to obtain a FQDN, and only hitting nodeDao once.

Refactor tests:
 - Clean up naming. Test classes are now prefixed with 'Test', methods with 'test'.
 - Wire in a mockNodeDao
 - Add tests for the 3 node lookup strategies
 - Add asserts for correctly indexed event severity.

NMS-8704: Minor cleanup

NMS-8704: Add documentation and default configuration.

NMS-8704: Fix typos in event definition and admin documentation, and don't set the description in the parser.