NMS-8650: Temporary fix CoreWeb to fix compilation issues.
There are other active GWT app not related with the Assets app, that
requires some dependencies declared inside the Asset app. As soon as the
Asset app was removed, those dependencies were gone and of course
opennms-webapp was broken.

I added the dependencies to CoreWeb in the meantime to be sure that
opennms-webapp compiles.