HZN-852: Separated Styles and neutralized trend chart layout
Change styling to a neutral layout. The color "green" is often interpreted as "OK" in monitoring. The charts itself
don't have any logic for such a thing. The hard coded CSS styling in JSP pages is broken out into SCSS files and use as
most possible the color variables from SCSS property files. I did also minor layout fixes for spacings, font size and
possitioning of the chart widgets.

The new chart styling is also easier to be used in Horizon/Meridian releases and don't require changes in the layout and
can be used in both products without forking the stylesheets.