Backport of Alejandro's Fix for Bug NMS-8779
--- Original Commit Text ---
Installer script doesn't work with PostgreSQL 9.6

The "CREATE USER" statement used to have "CREATEDB CREATEUSER".
CREATEUSER has to be replaced with CREATEROLE.

Now, neither of those flags are required for the opennms user because
the database is created by the provided superuser, as you can see in
the Migrator class. For this reason, they have been removed. OpenNMS
works as expected.

The JDBC driver has been updated to avoid issues with the installer.