Minor improvement added to MagicUsersMigratorOffline to avoid confuse users.
The initial implementation, used to show an error message when the
magic-users.properties file is not found while constructing the task
instance. This can confuse users, as the task will always be
constructed, but the execution won't be performed if it was executed
before. The error in question should be shown only if it is executed
(during the execution phase) and the file doesn't exist.

I made this change based on my upgrade tests.