Fix for Bug NMS-8879 - Part II
Replacing save() with saveOrUpdate() in case the asset record that is
supposed to be updated doesn't exist and it should be created (which is
technically impossible, but it doesn't hurt to have this just in case).
It is impossible because a node added by Provisiond (no matter if it is
auto-discovered or not) will have an asset entry on the assets table.