NMS-8085: Use platform independen resource path
The NIO Path library was used to identify a resource. These paths are
platform dependend and underly the rules of the OS running on. As
resource path can contain arbitary element names, the API was changed to
use ResourcePath instead.

This also replaces collons in the path with underscores to fix the
original problem of storing IPv6 addresses on windows.