NMS-9078: auto-detect best available Pinger API
Up until now, there has been only one concrete implementation of the PingerFactory
(PingerFactoryImpl), but this adds a replacement, BestMatchPingerFactory,
which attempts to figure out the best available Pinger to respond with.  This patch
also provides a backwards compatible Pinger implementation (BestMatchPinger) with
the same behavior that can be set directly as the default instance where appropriate.