NMS-9278: Added CriteriaBehavior + CXF SearchBean to RESTv2
Added ability to use the CXF SearchBean instead of the entity
object as the query capture bean. Added a CriteriaBehavior object that
can map query terms to new names, conversions, and actions. Added a common
set of aliases and behaviors that can be reused in different endpoints
when you are joining the same aliases.