HZN-1061: Add RESTv2 endpoint for alarms (#1567)
* NMS-9278: Initial work.

Basic end-points added.
The solution can be refactored.
Some tests are still failing.

* NMS-9278: Fixing errors on JUnit tests.

* NMS-9278: Refactoring AbstractDaoRestService to simplify implementations.

* NMS-9278: More refactoring.

Now, by default, AbstractDaoRestService will reject any POST, PUT and
DELETE. It is up to the class that extends this abstract class to
override the appropriate methods to provide options for create, modify
and delete entities.

* NMS-9278: More refactoring.

Fixing PUT logic.
Adding skeletons for JUnit tests (events, alarms).
Fixing end-point for alarms.

* NMS-9278: Refactoring update logic (this time it works).

* NMS-9278: Several improvements and adding IfServiceRestService.

* NMS-9278: Adding proper support for node categories and hardware inventory.

* NMS-9278: Adding location support for new end-points.

* NMS-9278: fixing compile validation.

* NMS-9278: Several improvements

- Creating a helper class for security checks, to have a common place
  for this kind of things.
- Adding action updates for the alarms endpoint.
- Adding ability to change sticky/journal memos for alarms.
- Update AbstractDaoRestService to provide the security context.

* NMS-9278: Improving support for JSON; Fixing POST events.

JSON support requires more work.

* NMS-9278: Upgrading CXF and fixing the nested object filtering capabilities

CXF was upgraded because of the following issue:
https://issues.apache.org/jira/browse/CXF-7213

* NMS-9278: Fixing several issues found during runtime tests.

1. CriteriaBuilderSearchVisitor was not working properly when specifying
multiple filters.

2. The distinct operation is required on several places to avoid
duplicates. All the end-points have been updaged.

3. Added some tests for the events end-point.

* NMS-9278: Make InetAddress fields mappable and allow searching alarms by service.name and ipAddr string

* NMS-9278: Extend AlarmRestService to allow create/update/delete tickets

* NMS-9278: Minor cleanup in RESTv2 classes

* NMS-9278: Removed most distinct() Criteria directives

Added TODOs for places that we need join conditions.

* NMS-9278: Fixed NodeRestServiceIT

* NMS-9278: Added orderBy tests for alarm and event services

* Expanded Criteria API to support Restrictions.eqProperty()

Added support for setting a root alias on the Criteria.

* NMS-9278: Updated join conditions with root aliases

* NMS-9278: Cleaned up alias JOINs and added assetRecord aliases

* NMS-9278: Added CriteriaBehavior + CXF SearchBean to RESTv2

Added ability to use the CXF SearchBean instead of the entity
object as the query capture bean. Added a CriteriaBehavior object that
can map query terms to new names, conversions, and actions. Added a common
set of aliases and behaviors that can be reused in different endpoints
when you are joining the same aliases.

* NMS-9278: Added SQL restriction for category id and name

This works around a bug in our version of Hibernate (documented
in issue NMS-9470) when adding an alias with a join condition
on a many-to-many relationship (like node-to-category).

* NMS-9278: Added distPoller alias to EventRestService

* NMS-9278: Added IpLikeCriteriaBehavior

Enables iplike criteria against root entities of a REST service.

* NMS-9278: More work on javadoc and making the RESTv2 services consistent

* NMS-9278: Added README file for RESTv2 services

Minor cleanup in services.

* NMS-9278: Added more orderBy tests to AlarmRestServiceIT

Added root alias iplike behaviors.

* NMS-9278: Use parameterized CriteriaBuilder.sql()

* NMS-9278: Finished criteria behaviors for notifications and outages

Pasted lots of test stanzas into the tests for events, notifications
and outages.

* NMS-9278: Fix failing v2 rest service tests

* NMS-9278: Fix iplike search for topology

* NMS-9278: Implement DELETE on MonitoringLocationRestService and fix tests

* NMS-9278: Make BSM related ITs pass