NMS-9290: refactoring DroolsCorrelationEngineBuilder to support hot-add and hot-delete
The spring initialization was modified in order to support the on-demand
addition of new engine directories, and the on-demand removal of engine
directories. For this purpose, a reloadDaemonConfig is required, using
'DroolsCorrelationEngine' for the 'daemonName'.

To update an existing engine, the intial design will be used, which is,
use 'DroolsCorrelationEngine-XXX' where XXX is the name of the engine
(or RuleSet) for the daemonName used on the reloadDaemonConfig event.