NMS-9691: Implement the Sink pattern using AWS SQS (#1719)
* Adding a working version of the Sink Pattern using AWS SQS.

This requires AWS Credentials when used outside a VPC.
The JUnit test requires AWS Credentials, so it is disabled until the
AwsUtils class is re-worked to allow mocks for tests.

* NMS-9691 - Applying suggestions : Part I

- Declaring AWS versions on the main POM.
- Creating a dependencies project for AWS.
- Simplifying the Karaf feature for AWS.
- Adding initial draft for the documentation.

* NMS-9691 - Applying suggestions : Part II

- Fixing typo on features.xml
- Removing amazon-sqs-java-messaging-lib as a dependency
- Refactoring the implementation to use pure AWS-SDK calls for SQS

Now, all the code should be thread-safe.

* NMS-9691 - Applying suggestions : Part III

- Replacing class name prefix from 'Aws' to 'AmazonSQS'

* NMS-9691 - Applying AWS workflow suggestion.

- Removing the message after consuming it.

* NMS-9691 - Applying some enhancements

- Upgraded the AWS SDK to the latest version.
- Provide a way to tune the SQS settings when creating queues.
- Be able to create queues from either OpenNMS or Minion.
- Adding documentation for the custom parameters.

* NMS-9691 - Improving AmazonSQSUtils

* NMS-9691 - Applying suggestions : Part IV

* NMS-9691 - Improving/Fixing documentation.

* NMS-9691 - Reducing calls to getQueueUrl.

- Calling this function requires an API call. All the traffic against
  AWS should be minimized when possible.

* NMS-9691 - Introducing aws_queue_name_prefix

This is an optional property. When it is defined, a prefix will be
enforced on the queue name, to avoid collisions on SQS, when deploying
multiple OpenNMS environments on the same AWS Region.

* NMS-9691 - Applying suggestions : Part V

- Replacing AmazonSQSUtils with AmazonSQSManager
- The default implementation of AmazonSQSManager uses AWS, while there
  is a mock implemenation for testing purposes.

* NMS-9691 - Fixing compilation issue

* NMS-9691 - fixing the prefix logic

* NMS-9691 - Minor improvement on the credentials management

* NMS-9691 - Applying suggestions : Part VI

I made some tests on a Minion to simulate issues while connecting to
AWS, or to be more precise "sqs.${REGION}.amazonaws.com".

I found that the exception thrown is SdkClientException. Now, the cause
behind this exception is either UnknownHostException (DNS issues) or
SocketException (rechability issues).

The code will enter into a blocking infinite loop until the
communication is restored or a different exception is thrown,
waiting 5 seconds before attempt again.

I haven't tested the lopo yet (will do this next).

* NMS-9691 - Improving logic for send/receive messages.

- The error handling for connections has been improved and tested.
- Log messages have been improved to be more verbose, in order to have a
  clear idea about what's happening at OpenNMS level, and at Minion
  level.
- Upon reception, messages will always be deleted, regardless if they
  were dispatched. The reason for this, is to avoid having a message
  around if a wrong message has sent to the wrong queue.

* NMS-9691 - Applying suggestions : Part VI

Adding support for FIFO Queues.

- All the send/receive logic has been moved to AmazonSQSManager.
- Now the producers and consumers are agnostic about the queue type, and
  this is handled by the implemenation of the AmazonSQSManager
  interface.
- FIFO queues are optional, so the user should enabled them on the
  configuration files (on both sides, OpenNMS and Minion).
- The documenation has been updated.

Other changes:

- AmazonSQSDispatcherBlueprintIT as been re-enabled, as it should not
  need to access AWS (it just verifies that the blueprint is loadable).

* NMS-9691 - Minor update to the docs.

* NMS-9691 - Fixing POMs after rebase to origin/release-21.0.1

* NMS-9691 - Improving documentation with AWS Credential details.

Using the static credentials provider, only when the keys are provided
through the configurtion file. Otherwise, use the default provider
chain.

* NMS-9691 - Improving documentation with AWS Credential details (Part II)

When running OpenNMS on an EC2 instance, you can assign an IAM Role with
permissions to have full access to SQS, as an alternative to hardcode
the AWS credentials on a configuration file.

* NMS-9691 - Applying suggestions : Part VII

Renaming the Sink strategy from 'aws' to 'sqs', and updating all related
configuration files.

Moving queue default parameters to DefaultAmazonSQSManager, and
improving the logic to build the runtime queue parameters (verified with
a JUnit test).

* NMS-9691 - Fixing docs for FIFO queues.

* NMS-9691 - Making sure the queue name is correct when FIFO is enabled.