HZN-1323: Update alarmd to use Drools to manage the alarms, replacing the existing Vacuumd automations.
Introduce the AlarmEntityNotifier interface to help capture
changes/updates to alarms and allow services to get callbacks
when these are performed.

Refactor the alarm nortbounder management to the NorthbounderManager class
and take advantage of the alarm entity change callbacks to trigger these.

Enhance the AlarmLifecycleListener interface to provide periodic
synchronization callbacks, and update the current listener implementations
to take advantage of this.