fix karaf-maven-plugin execution
Some of the new projects refactored as part of the Sentinel changes
included un-versioned plugin references to karaf-maven-plugin.  4.2.1
has a regression compared to 4.2.0 that caused our build to crash,
and it has just gotten pushed to the maven repos today so builds
started failing.

Also, I changed the shared container bootstrap back to the way
it used to be now that I know the differences were for 4.2.0 and
documented the new way if we ever bump karaf.