Improved logic for triggering ifOperStatus events
Triggering ifOperStatus events regardless of whether or ifAdminStatus
changed proved to be noisy and less than ideal.

With this commit, the behaviour is more similar to the logic before
suppressAdminDownEvent was added back to the code.

If the status change is because of ifAdminStatus:
- ifOperStatus DOWN will *only* be sent when ifAdminStatus goes UP
  while ifOperStatus remains DOWN.
- All ifAdminStatus events will be suppressed when suppressAdminDownEvent
  is enabled. The UP events are fairly useless without corresponding
  DOWN events.

If the status change is *not* related to a change in ifAdminStatus:
- ifOperStatus events will be sent as normal, unaffected by any
  of the changes related to re-adding suppressAdminDownEvent.