Add core pool size to MBeans and improve the graphs
Instead of graphing the max pool size, which is never used unless
the queue is full, include the current pool size and the core pool
size.

Also replaces peak active with current pool size. With the
current implementation, the current pool size seems to never shrink,
so the peak pool size would be redundant, otherwise it might make
sense to graph both metrics.