HZN-1579: Publish Docker images to DockerHub for master and develop branch
Created a little python script which fetches the version number from the root pom.xml which is used as OCI tags for
Horizon, Minion and Sentinel. Added a filter for the develop and master branch so they are only published to DockerHub.
When we push changes to master the version number from the root pom.xml is used, for develop we use a floating tag
"bleeding".