NMS-12165: Meta-Data works inside arbitrary values
This enables support for meta-data DSL expressions inside of all poller
parameters, whether they are defined in the value attribute or in the
body of the parameter tag.

If the body contains a nested XML structure, the structure is allowed to
contain expressions inside every attribute and element body. Validation
of the nested strucutre takes place after the expressions are evaluated
and have been replaced with the final value.

The Poller API was changed to use the PollerParameter interface instead
of arbitrary objects in the parameter map. While introducing the poller
parameter interface to the poller API, the detectors and data
collections have been altered to use only strings.