NMS-12165: Meta-Data works inside arbitrary values
This enables support for meta-data DSL expressions inside of all poller
parameters, whether they are defined in the value attribute or in the
body of the parameter tag.

If the body contains a nested XML structure, the structure is allowed to
contain expressions inside every attribute and element body. Validation
of the nested strucutre takes place after the expressions are evaluated
and have been replaced with the final value.

The Poller configuration was changed to defer unmarshalling of complex
structures and preserves the original XML values. The poller
implementations are now responsible for unmarshalling the complex values
into the concrete types. While simplifying the pollers, the detectors
and data collections have been altered to use only strings.