NMS-12211: Set format to PEM for SSH host key generation for Karaf
With OpenSSH 7.8+ the default format for RSA key pairs has been changed and writes by default the OpenSSH format. This
file can't be processed correctly from our current Karaf version. The key needs to be generated with the old PEM format
which is here enforced by adding the "-m PEM" as arguments.