Fix for Bug NMS-12360 - Part II
When requesting the requisition statistics for a given requisition,
the end-point will return an error if the requisition doesn't exist.
That is not the case when requesting all the stats, as that would
return an empty array.

Because the requisition must be updated with the statistics only when
they are available, the code was slightly modified to silently fail
when there are errors retrieving the stats, meaning the code can
continue with the requisition it already has preventing side effects.