NMS-12430: Use netty buffer instead of NIO buffer
This changes the packet parsing of ALL protocol implementations to use
nettys ByteBuf instead of the NIO ByteBuffer. This has the purpose of
avoiding  to copy the payload buffer and thus marking the whole content
as consumed. Instead, if a parser does not consume the whole content, it
will be passed to the parser again if new data arrives.