NMS-13324: Store service monitor status in RRD
The adds a flag `rrd-status` to service monitor definitions allowing to
store the status of the monitor in RRD (or other performance data
repository). The status is stored in a new repository directory called
`status` which follwos the same semantics as the `response` repository.

To query the monitor status, a new resource type is added allowing to
query the current status and the status history using the measurements
API.

The flag is disabled by default and must be enabled for each monitor.

To clarify things in UI, the labels for the status and response time
resources have been modified to show the difference.