NMS-13314: Prevent meta-data modification
Modification of meta-data was only intended to be done via requisitions.
The requisition process will create, update and delete meta-data of a node
according to the values defined in a requisition. Changing the meta-data
after a requisition synchronization was never intended conceptually. It
suffered from loosing of meta-data after re-synchronisation. Keeping the
data would require manual deletion in massive scale using the UI and
leads to desync with the expected values.