NMS-13294: Refactoring duplication in repository description
Moved Horizon and Meridian RHEL/CentOS repository description into a dedicated file and include it in the installation
guides for Core, Minion and Sentinel. For the reason we have RHEL7 and RHEL8 repositories I had to create dedicated
files for each of them. We can now maintain just the files in repos directory.

[ci skip]