NMS-13433: Publish Minion Docker image after smoke tests
We want to avoid publishing Minion images to the registry which don't pass the Minion tests. I've added an additional
requirement to finish the Minion smoke tests first before building and publishing single and multi-arch Minion images.
The Minion OCI image which is used in the Minion smoke tests is built very early in the Minion tarball assembly step.