dependency updates, and changes to support them
We have a bunch of dependabot updates, not all of them are strictly
necessary as they are only "security" issues at compile-time, not
runtime.  However, there are a few bugs that affect actual
serialization from compile, so I've updated those and things related
to them, including:

* babel
* eslint
* moment-javaformat
* webpack and some related loaders/plugins
* yargs

Also a couple of minor version bumps that are strictly for code we
use at runtime:

* c3
* core-js
* cron-parser
* ip-address
* jquery-ui
* moment
* regenerator-runtime
* ui-bootstrap4
* underscore

Note that one side-effect of updating webpack is that we ran into
an Uglify bug that we were just lucky to have not run into
previously.  This forced me to (finally) update to the Terser
plugin, which is a drop-in Uglify replacement.