NMS-13945: Use current string attrs for threshold filters
While evaluating the thresholds resource-filter, the attribute have been
read from the RRD repository (as stored in strings.properties). This
will fail if the string is not persisted (because the collection set us
purely used for thresholding) or on the first thresholding run, if the
thresholding is executed before the persistence.

This is changed to use the values from the currently processed
collection set.