Improve docker instructions
With running as a non-root user we have to make sure the etc directory has permissions set to a user with user id 10001. As a safe default, I've introduced a persistent volume that has automatically the correct permissions. I've created a tip that explains that you need to make sure to set the correct permission in case you want to use a bind mount from your local filesystem into the container.