create "core" and "shared" feature files
One of the issues I'm running into while trying to upgrade guava is that
on Minion and Sentinel, we want "real" dependencies so things get
loaded, but in the Horizon core, we want to do nothing because we're
exporting those dependencies from the `lib/` dir and the dependencies
are already present because of `custom.properties`.

It is always necessary to use the `custom.properties` version on core if
we are sharing code across the spring<->OSGi bridge or we run into
classloader issues.

This change provides a way to "depend" on something, while still
assuming it exists on the core side.