Upgrade to OpenTracing 0.32
This is in preparation for using the OpenTracing->OpenTelemetry shim,
doesn't work with 0.31 due to changes in the API. The only changes
that appear to directly impact us are the TextMap and related changes,
although there are deprecations that will impact us in the future.

I first tried upgrading to 0.33, however the 0.32 deprecations are
removed in 0.33, and this would require more involved code changes
and the OpenTracing->OpenTelemetry shim appears to work just fine
with 0.32.

Changelog: https://github.com/opentracing/opentracing-java/blob/master/CHANGELOG.md