Fix itests that break when they don't have an Identity configured
I don't know why these don't fail all the time, but oh well.

There is a lot of duplication in anonymous implementations of
Identity and MockIdentity/MockMinionIdentity that could stand
to be deduplicated into a single MockIdentity at some point,
but they're all over the place, so it's slightly non-trivial.