When we fail to startup, exit with a non-zero exit code.
Docker/Kubernetes don't seem to do what we'd want when a container
for a service like OpenNMS stops running because the process exits 0.

Things take a somewhat circuitous path when stopping until doSystemExit
is called, so we stash the exit code on a new field so doSystemExit has
access to it a few later (and a few layers deeper).

While I'm here, log when we are shutting down due to a SIGTERM (but
still exit 0 in this case, I guess?).