fix(oci): Remove deprecated apt-key method in container build
Replace the apt-key deprecated way adding a GPG key for signed packages on a Debian package repository. Instead of
adding the signing-key to the global apt repository, which is applied to any package for all repositories, we add a
specific keyring file and use it just for repository described in the opennms.list file.

Resolves: NMS-14760