Enable recording when there is a test failure
Add support for using alternate Docker images in
OpenNMSImageNameSubstitutor. It is enabled by configuration in the
testcontainers.properties file. Use a system property to get the
seleniarm version.

Work around an issue where org.testcontainers doesn't stop recording
when the test is done.

Since we use BrowserWebDriverContainer as a @ClassRule, not a @Rule,
add a @Rule to keep track when tests fail so we can properly tell
BrowserWebDriverContainer.afterTest if there were failures. This lets
it properly decide whether to save the recording.

Use mp4 since pretty much everything supports it well.