Add OpenNMSStack.NONE for local dev with components outside of Docker
Use this when you want to test locally in your dev environment using
local PostgreSQL, OpenNMS, etc. instead of standing up components
inside of Docker. This is useful when you are narrowing down a
problem or trying fixes and won't want to incur the overhead of
starting OpenNMS in a container every time, for example.