Make awaitHealthCheckSucceeded pass/fail quickly in common cases
Test immediately so it can pass quickly if the Karaf container is up.

Fail quickly if the container is no longer running.

Pass the container to awaitHealthCheckSucceeded so the latter check
can be done and add getSshAddress to KarafContainer so we can drop
the rest of the arguments.

Since we no longer pass OpenNMS or Minion, include the Docker image
name so we know what was tested for the health check.