I don't think we need log karaf to the console anymore
With the changes to exception handling and better debugging
information there, I'm not sure if makes sense to log karaf to
the console anymore. It will just end up in output.log now,
anyway, so might as well go look at karaf.log if you need to
see it.

This would be a good one to sanity check with someone who does
debugging in these parts, though.