Speedup thread dump collection and store in its own file
Instead of always sleeping 5 seconds, wait for the end of the thread
dump to be stored. In my experience, this takes about 300 ms.

While we're here, since we know where the thread dump is collected
and when it is done, stash it in its own file This makes it super
easy to open up the thread dump in IntelliJ.  We still have a copy
in the console logs / output.log, unfortunately.

This also required separating out when we cleanup the existing log
directory, so we don't gather the thread dump and then immediately
delete it.

Inline copyLogs in the Container classes, since it was only being
used once and I need to get at some of the variables in it.

While we're here, log the path to the common logs for our main
three container classes like we do for Horizon today.