NMS-15001: Use service locator key in poller client
The lookup of the service monitor implementation can not be done by
class name, as there are delegate implementaions like the one used for
OIA pollers. Therefor the pollers are now addressed by service locators
only. This uses the registered poller class name, which can differ from
the actual class in case of delegates. Thus allowing to find the correct
implementations in the service monitor registry.