NMS-15100: Change needed to a file name reference in documentation
NMS-15100: Change needed to a file name reference in documentation - updated the name of the file that needs to be moved as part of the TSDB set up.
NMS-14820: Documentation doesn't render asciimath properly: removed the 3 formulas that were listed. They did not add any value, and the work involved to add an asciimath renderer to the environment outweighed any benefit.