Create JaegerContainer for smoke tests
Change JaegerTracingIT to use built-in RestAssured JSON path assertions,
which greatly simplifies the tests and improves error reporting. I just
had to get the right JSON path syntax. ;-)

Also try to quiet down logging on retried REST calls. If this works well,
it could stand to be moved to a helper somewhere.