Add container labels to a file inside of the Horizon container
This creates the file /usr/share/opennms/.container-labels with
the same key/value pairs that are passed to LABEL in Dockerfile
(with some empty/unset values removed). This makes the build
metadata available for debugging when looking at a container
image and for future use to help with observability tools like
OpenTelemetry for resource attributes.

There are a few things to consider along with this change:
1. Apply the same to all containers, not just Horizon.
2. Consider eliminating the LABEL commands in the Dockerfiles
   and passing in the labels on the command-line. This would
   have the implication that builds would only need to be done
   with the Makefile. Note: this would also allow us to de-
   duplicate other things like BASE_IMAGE that are currently
   set in both common.mk and in each Dockerfile.
3. Consider storing the labels in the container as a JSON file
   so it can be easily parsed with modern tools like jq.

Other things noticed: we might not want to put the version
number in the title. It might also be useful to put the plain
product name in a field, as well.

Random thought: should we have CLI tool inside bin that can
understand which version of OpenNMS is running, regardless of
whether it is packaged into an RPM, Debian package, or container?