Add overlay file support to MinionContainer
This is to support a future smoke test for Minion container overlay
file support

Unify the MinionContainer constructors and require a MinionProfile.

Add support for legacy configuration through environment variables to
MinionProfile (withLegacyConfiguration).

For legacy Minion containers, OpenNMSStack now creates a MinionProfile.

Eliminated a stray "container" field on MinionContainer that was just
a reference to the object.