"/." all the things! (well, copy target directories, at least)
This enforces that the target directory exists or the command fails.

It's also a good habit to get into with 'cp', when doing recursive
copies of directories in particular, because you will get different
results if you just do a target of "dir" or "dir/" (without the
trailing period) depending on whether or not the directory exists.
And that's a fun bug to chase down when it's not clear why it is
acting differently in some cases.

BTW, "dir", "dir/" or "dir/." doesn't seem to matter for rsync if the
source is a directory or if there is more than one source--it will
create the target directory if it doesn't exist. But not if the source
is a single file and the destination is "dir".