Suggested by Bonrob1
Note in line 14, it should be OSPF topology updater.

To align with Nant style guidelines, we use initial caps only for the first heading in a chapter. Subsequent headings use sentence case where only the first letter of the heading is capital, unless a word is a proper noun or an acronym.

Also, in line 18, "Node A" should be "node A"