EXPERIMENTAL: only wait for uptime >= grace period
This is to assist with testing code changes to prevent every call
to get() while the timeseries feature isn't active from taking 5 minutes.
Instead, it will only wait for a 5 minute grace period from when the JVM
started.

I have no idea if this is the right fix or not.

I really don't know what I'm doing here.