NMS-15413: fix for CortexTSS tests flapping with retries
It seems that the rate of flapping seems to have increased significantly
after the KAR files were moved into the container.

These tests haven't been changed, so they were downloading the Cortex
KAR file and copying it into the container. I suspect that Karaf takes
longer to deactivate and activate KAR files when this happens. I looked
for a few logs in CI, but ran into NMS-15427 and couldn't readily find
logs to see if this is what is happening.