NMS-16406: Allow update-package-permissions to set ownership for customized users.
Because the step that changes file ownership keys off of the RUNAS
variable, it does not make sense to bail out when RUNAS does not equal
`opennms` and it can leave the installation in an inconsistent state.

Modify the message printed to indicate that file ownership will be
changed to the RUNAS user, and do not exit immediately.