feat: Add quick-build short cut and add 32GB RAM build for Sentinel e2e tests
The make quick-compile && make quick-assemble is a common task. Having a target quick-build
can make things a bit easier for developers. You can just check out the repository and run make
which uses the default target quick-build.
Use the ionice command and make sure it is only used when it is available.

Added a label for the Sentinel E2E tests who require a machine with 32GB RAM to pass. Added a manual dispatch workflow
so users can trigger a build without the need to push somthing to a branch.

Add smoke test logs and screentshots to test results