refactor: Unify PostgreSQL and Core service in a docker-compose.yml
The reference to a stack of services in a single docker-compose.yml file is easier to follow.
We have the database name as a resolvable service name which can be easily used in the core service.
The change provides a single copy & paste stack with the database configuration and removes friction.

Signed-off-by: Ronny Trommer <ronny@no42.org>