NMS-18017: make expression thresholds more human-readable (#7749)
* This adds a new attribute for thresholds named "expr-label" and exposes
it in the webui threshold editor for expression-based thresholds. The intention of this attribute is to
allow for a "friendly" name or descriptor for what could be a
complicated threshold expression; when displayed in an alarm logmsg "((EnqueCnt + DequeCnt) == 0) ? 0 : (((EnqueCnt -
DequeCnt) / ((EnqueCnt + DequeCnt) / 2)) * 100)" can be
incomprehensible, whereas an expression label "ActiveMQ Queue Throughput
Mismatch" may be clearer to an operator.

* fix rrdrepo path

* testing artifact

* Clarify the purpose fo expressionsn label versus decription

* Missed in cleanup

* Per review; add field to docs